refactor: the path not updated in previous commits
